+++ dart/lib/src/protocol/sentry_span.dart
@@ -19,17 +19,17 @@ class SentrySpan extends ISentrySpan {
 
   SpanStatus? _status;
   final Map<String, String> _tags = {};
-  void Function({DateTime? endTimestamp})? _finishedCallback;
+  Function({DateTime? endTimestamp})? _finishedCallback;
 
   @override
-  bool? sampled;
+  final SentryTracesSamplingDecision? samplingDecision;
 
   SentrySpan(
     this._tracer,
     this._context,
     this._hub, {
     DateTime? startTimestamp,
-    this.sampled,
+    this.samplingDecision,
     Function({DateTime? endTimestamp})? finishedCallback,
   }) {
     _startTimestamp = startTimestamp?.toUtc() ?? getUtcDateTime();
@@ -62,7 +62,7 @@ class SentrySpan extends ISentrySpan {
     if (_throwable != null) {
       _hub.setSpanContext(_throwable, this, _tracer.name);
     }
-    _finishedCallback?.call(endTimestamp: _endTimestamp);
+    await _finishedCallback?.call(endTimestamp: _endTimestamp);
     return super.finish(status: status, endTimestamp: _endTimestamp);
   }
 
@@ -180,6 +180,21 @@ class SentrySpan extends ISentrySpan {
   SentryTraceHeader toSentryTrace() => SentryTraceHeader(
         _context.traceId,
         _context.spanId,
-        sampled: sampled,
+        sampled: samplingDecision?.sampled,
       );
+
+  @override
+  void setMeasurement(
+    String name,
+    num value, {
+    SentryMeasurementUnit? unit,
+  }) {
+    _tracer.setMeasurement(name, value, unit: unit);
+  }
+
+  @override
+  SentryBaggageHeader? toBaggageHeader() => _tracer.toBaggageHeader();
+
+  @override
+  SentryTraceContextHeader? traceContext() => _tracer.traceContext();
 }