+++ dart/test/sentry_client_test.dart
@@ -781,6 +781,21 @@ void main() {
       );
       expect(event.fingerprint!.contains('process'), true);
     });
+
+    test('thrown error is handled', () async {
+      final exception = Exception("before send exception");
+      final beforeSendCallback = (SentryEvent event, {dynamic hint}) {
+        throw exception;
+      };
+
+      final client =
+          fixture.getSut(beforeSend: beforeSendCallback, debug: true);
+
+      await client.captureEvent(fakeEvent);
+
+      expect(fixture.loggedException, exception);
+      expect(fixture.loggedLevel, SentryLevel.error);
+    });
   });
 
   group('EventProcessors', () {
@@ -1066,6 +1081,35 @@ void main() {
       expect(envelope.clientReport, clientReport);
     });
 
+    test('captureEvent adds trace context', () async {
+      final client = fixture.getSut();
+
+      final scope = Scope(fixture.options);
+      scope.span =
+          SentrySpan(fixture.tracer, fixture.tracer.context, MockHub());
+
+      await client.captureEvent(fakeEvent, scope: scope);
+
+      final envelope = fixture.transport.envelopes.first;
+      expect(envelope.header.traceContext, isNotNull);
+    });
+
+    test('captureTransaction adds trace context', () async {
+      final client = fixture.getSut();
+
+      final tr = SentryTransaction(fixture.tracer);
+
+      final context = SentryTraceContextHeader.fromJson(<String, dynamic>{
+        'trace_id': '${tr.eventId}',
+        'public_key': '123',
+      });
+
+      await client.captureTransaction(tr, traceContext: context);
+
+      final envelope = fixture.transport.envelopes.first;
+      expect(envelope.header.traceContext, isNotNull);
+    });
+
     test('captureUserFeedback calls flush', () async {
       final client = fixture.getSut(eventProcessor: DropAllEventProcessor());
 
@@ -1206,6 +1250,9 @@ class Fixture {
   late SentryTransactionContext _context;
   late SentryTracer tracer;
 
+  SentryLevel? loggedLevel;
+  Object? loggedException;
+
   SentryClient getSut({
     bool sendDefaultPii = false,
     bool attachStacktrace = true,
@@ -1214,6 +1261,7 @@ class Fixture {
     BeforeSendCallback? beforeSend,
     EventProcessor? eventProcessor,
     bool provideMockRecorder = true,
+    bool debug = false,
   }) {
     final hub = Hub(options);
     _context = SentryTransactionContext(
@@ -1228,6 +1276,8 @@ class Fixture {
     options.attachThreads = attachThreads;
     options.sampleRate = sampleRate;
     options.beforeSend = beforeSend;
+    options.debug = debug;
+    options.logger = mockLogger;
 
     if (eventProcessor != null) {
       options.addEventProcessor(eventProcessor);
@@ -1244,5 +1294,16 @@ class Fixture {
   FutureOr<SentryEvent?> droppingBeforeSend(SentryEvent event,
       {dynamic hint}) async {
     return null;
+  }
+
+  void mockLogger(
+    SentryLevel level,
+    String message, {
+    String? logger,
+    Object? exception,
+    StackTrace? stackTrace,
+  }) {
+    loggedLevel = level;
+    loggedException = exception;
   }
 }