+++ dart/test/sentry_traces_sampler_test.dart
@@ -13,11 +13,11 @@ void main() {
     final trContext = SentryTransactionContext(
       'name',
       'op',
-      sampled: true,
+      samplingDecision: SentryTracesSamplingDecision(true),
     );
     final context = SentrySamplingContext(trContext, {});
 
-    expect(sut.sample(context), true);
+    expect(sut.sample(context).sampled, true);
   });
 
   test('options has sampler', () {
@@ -36,7 +36,7 @@ void main() {
     );
     final context = SentrySamplingContext(trContext, {});
 
-    expect(sut.sample(context), true);
+    expect(sut.sample(context).sampled, true);
   });
 
   test('transactionContext has parentSampled', () {
@@ -45,11 +45,11 @@ void main() {
     final trContext = SentryTransactionContext(
       'name',
       'op',
-      parentSampled: true,
+      parentSamplingDecision: SentryTracesSamplingDecision(true),
     );
     final context = SentrySamplingContext(trContext, {});
 
-    expect(sut.sample(context), true);
+    expect(sut.sample(context).sampled, true);
   });
 
   test('options has rate 1.0', () {
@@ -61,7 +61,7 @@ void main() {
     );
     final context = SentrySamplingContext(trContext, {});
 
-    expect(sut.sample(context), true);
+    expect(sut.sample(context).sampled, true);
   });
 
   test('options has rate 0.0', () {
@@ -73,19 +73,57 @@ void main() {
     );
     final context = SentrySamplingContext(trContext, {});
 
-    expect(sut.sample(context), false);
+    expect(sut.sample(context).sampled, false);
+  });
+
+  test('tracesSampler exception is handled', () {
+    final sut = fixture.getSut(debug: true);
+
+    final exception = Exception("tracesSampler exception");
+    double? sampler(SentrySamplingContext samplingContext) {
+      throw exception;
+    }
+
+    fixture.options.tracesSampler = sampler;
+
+    final trContext = SentryTransactionContext(
+      'name',
+      'op',
+    );
+    final context = SentrySamplingContext(trContext, {});
+    sut.sample(context);
+
+    expect(fixture.loggedException, exception);
+    expect(fixture.loggedLevel, SentryLevel.error);
   });
 }
 
 class Fixture {
   final options = SentryOptions(dsn: fakeDsn);
 
+  SentryLevel? loggedLevel;
+  Object? loggedException;
+
   SentryTracesSampler getSut({
     double? tracesSampleRate = 1.0,
     TracesSamplerCallback? tracesSampler,
+    bool debug = false,
   }) {
     options.tracesSampleRate = tracesSampleRate;
     options.tracesSampler = tracesSampler;
+    options.debug = debug;
+    options.logger = mockLogger;
     return SentryTracesSampler(options);
+  }
+
+  void mockLogger(
+    SentryLevel level,
+    String message, {
+    String? logger,
+    Object? exception,
+    StackTrace? stackTrace,
+  }) {
+    loggedLevel = level;
+    loggedException = exception;
   }
 }